comitted testNG emailable report.java and updated Base and log4j.properties and deleted customized emailable report.java
