+++ src/main/resources/properties/log4j.properties
@@ -1,20 +1,9 @@
 
 
 
-# Root logger option
-log4j.rootLogger=INFO, file, stdout
-
-# configuration to print into file
-log4j.appender.file=org.apache.log4j.RollingFileAppender
-#log4j.appender.file.File=C:\\Users\\Ideas2it\\Desktop\\ETQ_PRECHECK\\target\\+envId+Outputsystem.log
-log4j.appender.file.MaxFileSize=12MB
-log4j.appender.file.MaxBackupIndex=10
-log4j.appender.APP.Append = false
-log4j.appender.file.layout=org.apache.log4j.PatternLayout
-log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %p [%c{3}] [%M]: %m%n 
-
-# configuration to print on console
+log4j.rootLogger=INFO, stdout
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
 log4j.appender.stdout.Target=System.out
 log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{3}:%L - %m%n
+log4j.appender.stdout.layout.ConversionPattern=%d{yy/MM/dd HH:mm:ss} %p [%c{3}] [%M]: %m%n
+
